+++ R/analysis.R
@@ -3,7 +3,8 @@ Analysis <- function(Functions, condition, replications, fixed_objects, cl, MPI,
                      boot_method, boot_draws, CI, save_seeds, save_seeds_dirname, load_seed,
                      export_funs, summarise_asis, warnings_as_errors, progress, store_results,
                      allow_na, allow_nan, use_try, stop_on_fatal, store_warning_seeds,
-                     include_replication_index, packages, .options.mpi, useFuture)
+                     include_replication_index, packages, .options.mpi, useFuture, multirow,
+                     allow_gen_errors, save_results_filename = NULL)
 {
     # This defines the work-flow for the Monte Carlo simulation given the condition (row in Design)
     #  and number of replications desired
@@ -24,7 +25,8 @@ Analysis <- function(Functions, condition, replications, fixed_objects, cl, MPI,
                                                    warnings_as_errors=warnings_as_errors,
                                                    include_replication_index=include_replication_index,
                                                    allow_na=allow_na, allow_nan=allow_nan, use_try=use_try,
-                                                   p=p, future.seed=TRUE), silent=TRUE)
+                                                   p=p, future.seed=TRUE, allow_gen_errors=allow_gen_errors),
+                       silent=TRUE)
     } else if(is.null(cl)){
         if(!is.null(seed)) set.seed(seed[condition$ID])
         results <- if(progress){
@@ -39,7 +41,8 @@ Analysis <- function(Functions, condition, replications, fixed_objects, cl, MPI,
                    save_seeds_dirname=save_seeds_dirname,
                    warnings_as_errors=warnings_as_errors,
                    include_replication_index=include_replication_index,
-                   allow_na=allow_na, allow_nan=allow_nan, use_try=use_try), TRUE)
+                   allow_na=allow_na, allow_nan=allow_nan, use_try=use_try,
+                   allow_gen_errors=allow_gen_errors), TRUE)
         } else {
             try(lapply(1L:replications, mainsim, condition=condition,
                    generate=Functions$generate,
@@ -52,7 +55,8 @@ Analysis <- function(Functions, condition, replications, fixed_objects, cl, MPI,
                    save_seeds_dirname=save_seeds_dirname,
                    warnings_as_errors=warnings_as_errors,
                    include_replication_index=include_replication_index,
-                   allow_na=allow_na, allow_nan=allow_nan, use_try=use_try), TRUE)
+                   allow_na=allow_na, allow_nan=allow_nan, use_try=use_try,
+                   allow_gen_errors=allow_gen_errors), TRUE)
         }
     } else {
         if(MPI){
@@ -67,7 +71,7 @@ Analysis <- function(Functions, condition, replications, fixed_objects, cl, MPI,
                      store_warning_seeds=store_warning_seeds,
                      include_replication_index=include_replication_index,
                      warnings_as_errors=warnings_as_errors, allow_na=allow_na, allow_nan=allow_nan,
-                     use_try=use_try), TRUE)
+                     use_try=use_try, allow_gen_errors=allow_gen_errors), TRUE)
         } else {
             if(!is.null(seed)) parallel::clusterSetRNGStream(cl=cl, seed[condition$ID])
             results <- if(progress){
@@ -80,7 +84,8 @@ Analysis <- function(Functions, condition, replications, fixed_objects, cl, MPI,
                                     save_seeds=save_seeds, save_seeds_dirname=save_seeds_dirname,
                                     warnings_as_errors=warnings_as_errors, allow_na=allow_na,
                                     include_replication_index=include_replication_index,
-                                    allow_nan=allow_nan, use_try=use_try, cl=cl), TRUE)
+                                    allow_nan=allow_nan, allow_gen_errors=allow_gen_errors,
+                                    use_try=use_try, cl=cl), TRUE)
             } else {
                 try(parallel::parLapply(cl, 1L:replications, mainsim,
                                     condition=condition, generate=Functions$generate,
@@ -91,7 +96,8 @@ Analysis <- function(Functions, condition, replications, fixed_objects, cl, MPI,
                                     save_seeds=save_seeds, save_seeds_dirname=save_seeds_dirname,
                                     warnings_as_errors=warnings_as_errors, allow_na=allow_na,
                                     include_replication_index=include_replication_index,
-                                    allow_nan=allow_nan, use_try=use_try), TRUE)
+                                    allow_nan=allow_nan, allow_gen_errors=allow_gen_errors,
+                                    use_try=use_try), TRUE)
             }
         }
     }
@@ -109,10 +115,13 @@ Analysis <- function(Functions, condition, replications, fixed_objects, cl, MPI,
             return(ret)
         }
     }
+    ID <- ifelse(multirow, paste0('-', condition$ID), "")
     if(summarise_asis || store_results){
         tabled_results <- toTabledResults(results)
         if(save_results){
-            tmpfilename <- paste0(save_results_dirname, '/results-row-', condition$ID, '.rds')
+            tmp <- ifelse(is.null(save_results_filename), 'results-row', save_results_filename)
+            tmpfilename <- paste0(save_results_dirname,
+                                  sprintf('/%s', tmp), ID, '.rds')
             saveRDS(list(condition=condition, results=tabled_results),
                     file.path(save_results_out_rootdir, tmpfilename))
         }
@@ -139,7 +148,9 @@ Analysis <- function(Functions, condition, replications, fixed_objects, cl, MPI,
     #collect meta simulation statistics (bias, RMSE, type I errors, etc)
     results <- stackResults(results)
     if(save_results){
-        tmpfilename <- paste0(save_results_dirname, '/results-row-', condition$ID, '.rds')
+        tmp <- ifelse(is.null(save_results_filename), 'results-row', save_results_filename)
+        tmpfilename <- paste0(save_results_dirname,
+                              sprintf('/%s', tmp), ID, '.rds')
         tmpcondition <- condition
         tmpcondition$ID <- NULL
         saveRDS(list(condition=tmpcondition, results=results, errors=try_errors,
@@ -160,8 +171,9 @@ Analysis <- function(Functions, condition, replications, fixed_objects, cl, MPI,
     }
     sim_results <- sim_results_check(sim_results)
     summarise_list <- attr(sim_results, 'summarise_list')
-    ret <- c(sim_results, 'REPLICATIONS'=replications, 'ERROR: '=try_errors,
-             'WARNING: '=warnings)
+    ret <- c(sim_results, 'REPLICATIONS'=replications,
+             'ERROR: '=clip_names(try_errors),
+             'WARNING: '=clip_names(warnings))
     if(boot_method != 'none'){
         # could parallelize, but likely not worth the overhead
         # TODO test whether this works with Summarise() list outputs