+++ inst/doc/Parallel-computing.R
@@ -23,7 +23,7 @@ par(mar=c(3,3,1,1)+.1)
 #  IPs <- list(list(host=primary, user='myname', ncore=8), list(host='192.168.2.2', user='myname', ncore=6))
 #  spec <- lapply(IPs, function(IP) rep(list(list(host=IP$host, user=IP$user)), IP$ncore))
 #  spec <- unlist(spec, recursive=FALSE)
-#  cl <- makeCluster(master=primary, spec=spec)
+#  cl <- makeCluster(master=primary, spec=spec, type = 'PSOCK')
 #  Final <- runSimulation(..., cl=cl)
 #  stopCluster(cl)
 
@@ -33,15 +33,7 @@ par(mar=c(3,3,1,1)+.1)
 #  IPs <- list(list(host=primary, ncore=8), list(host='slave', ncore=6))
 #  spec <- lapply(IPs, function(IP) rep(list(list(host=IP$host)), IP$ncore))
 #  spec <- unlist(spec, recursive=FALSE)
-#  cl <- makeCluster(master=primary, spec=spec)
-#  Final <- runSimulation(..., cl=cl)
-#  stopCluster(cl)
-
-## ----eval=FALSE---------------------------------------------------------------
-#  library(parallel)
-#  primary <- 'master'
-#  spec <- c(rep(primary, 8), rep('slave', 6))
-#  cl <- makeCluster(master=primary, spec=spec)
+#  cl <- makeCluster(master=primary, spec=spec, type = 'PSOCK')
 #  Final <- runSimulation(..., cl=cl)
 #  stopCluster(cl)